+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Set errors property of `Spawnia\Sailor\Error\ResultErrorsException`
+
 ## v0.17.0
 
 ### Changed